Rename package containing NSI support
