+++ src/main/java/de/blau/android/App.java
@@ -25,10 +25,10 @@ import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import de.blau.android.filter.PresetFilter;
-import de.blau.android.names.Names;
-import de.blau.android.names.Names.NameAndTags;
 import de.blau.android.net.OkHttpTlsCompat;
 import de.blau.android.net.UserAgentInterceptor;
+import de.blau.android.nsi.Names;
+import de.blau.android.nsi.Names.NameAndTags;
 import de.blau.android.osm.DiscardedTags;
 import de.blau.android.osm.StorageDelegator;
 import de.blau.android.photos.Photo;
