+++ src/main/java/de/blau/android/nsi/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Classes for utilizing the OSM Name index for canonical store and similar chains
  */
-package de.blau.android.names;
+package de.blau.android.nsi;
