+++ src/main/java/de/blau/android/propertyeditor/TagEditorFragment.java
@@ -49,8 +49,8 @@ import de.blau.android.HelpViewer;
 import de.blau.android.R;
 import de.blau.android.address.Address;
 import de.blau.android.exception.UiStateException;
-import de.blau.android.names.Names;
-import de.blau.android.names.Names.NameAndTags;
+import de.blau.android.nsi.Names;
+import de.blau.android.nsi.Names.NameAndTags;
 import de.blau.android.osm.OsmElement;
 import de.blau.android.osm.OsmElement.ElementType;
 import de.blau.android.osm.Server;
