+++ src/main/java/de/blau/android/propertyeditor/tagform/TagFormFragment.java
@@ -45,8 +45,8 @@ import de.blau.android.App;
 import de.blau.android.HelpViewer;
 import de.blau.android.R;
 import de.blau.android.address.Address;
-import de.blau.android.names.Names;
-import de.blau.android.names.Names.NameAndTags;
+import de.blau.android.nsi.Names;
+import de.blau.android.nsi.Names.NameAndTags;
 import de.blau.android.osm.OsmElement;
 import de.blau.android.osm.Server;
 import de.blau.android.osm.Tags;
