+++ src/main/java/de/blau/android/propertyeditor/tagform/TextRow.java
@@ -14,8 +14,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import de.blau.android.App;
 import de.blau.android.R;
-import de.blau.android.names.Names;
-import de.blau.android.names.Names.NameAndTags;
+import de.blau.android.nsi.Names;
+import de.blau.android.nsi.Names.NameAndTags;
 import de.blau.android.osm.Tags;
 import de.blau.android.presets.Preset.PresetItem;
 import de.blau.android.presets.PresetComboField;
