+++ src/main/java/de/blau/android/util/SearchIndexUtils.java
@@ -19,8 +19,8 @@ import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import de.blau.android.App;
-import de.blau.android.names.Names.NameAndTags;
-import de.blau.android.names.Names.TagMap;
+import de.blau.android.nsi.Names.NameAndTags;
+import de.blau.android.nsi.Names.TagMap;
 import de.blau.android.osm.OsmElement.ElementType;
 import de.blau.android.prefs.Preferences;
 import de.blau.android.presets.Preset;
