+++ src/main/java/de/blau/android/voice/Commands.java
@@ -21,7 +21,7 @@ import de.blau.android.Main;
 import de.blau.android.R;
 import de.blau.android.address.Address;
 import de.blau.android.exception.OsmIllegalOperationException;
-import de.blau.android.names.Names.NameAndTags;
+import de.blau.android.nsi.Names.NameAndTags;
 import de.blau.android.osm.Node;
 import de.blau.android.osm.OsmElement.ElementType;
 import de.blau.android.osm.Tags;
